refactor(recall): resolve sentrux import fan-out violations and update task beans
