+++ .beans/pawrrtal-1cfl--active-recall-auto-search-lcm-before-the-main-agen.md
@@ -1,11 +1,11 @@
 ---
 # pawrrtal-1cfl
 title: 'Active Recall: auto-search LCM before the main agent runs'
-status: in-progress
+status: completed
 type: epic
 priority: high
 created_at: 2026-05-19T07:15:14Z
-updated_at: 2026-05-19T07:41:37Z
+updated_at: 2026-05-21T22:20:39Z
 ---
 
 ## What this is
@@ -30,3 +30,16 @@ Make the AI quietly look through old chat memory **before** it answers, so it "r
 - Lives at `backend/app/plugins/active_recall/`.
 
 > Inspired by OpenClaw's Active Memory.
+
+## Summary of Changes
+
+- Implemented `PreTurnHook` plugins architecture.
+- Added settings for active recall configuration (`active_recall_enabled`, `active_recall_model`, `active_recall_budget_seconds`, `active_recall_max_turns`).
+- Created the Active Recall plugin at `backend/app/plugins/active_recall/` consisting of:
+  - `plugin.py`: Registering pre-turn hooks.
+  - `recall_agent.py`: Implementing a lightweight sub-agent that searches long-term memory via LCM tools.
+- Integrated pre-turn hooks into `turn_runner.run_turn()`.
+- Threaded recalled context into the main agent's system prompt when available.
+- Added comprehensive unit and integration tests using `ScriptedStreamFn` for verifying sub-agent execution, tool limits, and timeouts.
+- Wrote ADR docs for Active Recall and updated the handbook.
+- Resolved Sentrux import fan-out architecture constraints.
